# apps/api/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Digilist Backoffice API
  description: |
    Complete API for the Digilist facility booking and management platform.

    ## Authentication

    Most endpoints require authentication via JWT token in the `Authorization` header:
    ```
    Authorization: Bearer <jwt-token>
    ```

    Additional headers for tenant context:
    - `X-Tenant-Id`: UUID of the tenant (required for most endpoints)
    - `X-License-Key`: License key for API access

    ## Rate Limiting

    API requests are limited to 1000 requests per minute per tenant.

    ## Response Format

    All responses follow a standard format:

    **Success (single resource):**
    ```json
    { "data": { ... } }
    ```

    **Success (list):**
    ```json
    { "data": [...], "meta": { "total": 100, "page": 1, "limit": 20, "totalPages": 5 } }
    ```

    **Error:**
    ```json
    { "error": { "code": "ERROR_CODE", "message": "Human readable message" } }
    ```
  version: 1.0.0
  contact:
    name: Xala Technologies
    email: api@xalatechnologies.com
    url: https://xalatechnologies.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.digilist.no
    description: Production server
  - url: http://localhost:4000
    description: Local development
...
paths:
...
  /api/auth/session:
    get:
      tags: [Auth]
      summary: Get current session
      operationId: getSession
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSessionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/authz/permissions:
    get:
      tags: [RBAC]
      summary: Get current user permissions
      operationId: getPermissions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User permissions
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/RolePermissions"
  /api/authz/check:
    get:
      tags: [RBAC]
      summary: Check if action is allowed
      operationId: checkPermission
      security:
        - bearerAuth: []
      parameters:
        - name: resource
          in: query
          required: true
          schema:
            type: string
          example: bookings
        - name: action
          in: query
          required: true
          schema:
            type: string
            enum: [read, create, update, delete]
          example: delete
      responses:
        "200":
          description: Permission check result
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      allowed:
                        type: boolean
                      role:
                        type: string
        "400":
          $ref: "#/components/responses/ValidationError"

  # ==========================================================================
  # Public Endpoints
  # ==========================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object

    # Auth Types
    AuthUser:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [super_admin, admin, saksbehandler, user]
    AuthSessionResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            user:
              $ref: "#/components/schemas/AuthUser"
            token:
              type: string
            expiresAt:
              type: string
              format: date-time
...
    RolePermissions:
      type: object
      properties:
        role:
          type: string
        permissions:
          type: object
          additionalProperties:
            type: array
            items:
              type: string

    # Listing Types
...
security:
  - bearerAuth: []